Require environment keys for cote to be set while running
The user has to give an environment key (such as "production", "test",
or the name of the machine) to make sure that multiple semaver instances
can run side by side.

The main problem here is that cote's auto-discovery works with broadcast
which in turn makes every daemon discover every other. This is a problem
in a shared environment where multiple semaver servers have to work
together. By giving out different keys to different servers, a user makes
sure that these servers won't communicate with each other.